v-contact-list: split VContactListModal into VContactList and VContactListModal
