+++ v-contact-list/v-contact-list-modal.css
@@ -1,9 +1,11 @@
-v-contact-list-modal {
+@import 'v-contact-list.css';
+
+.v-contact-list-modal {
     width: 100%;
 }
 
 @media (min-width: 540px) {
-    v-contact-list-modal {
+    .v-contact-list-modal {
         max-width: 534px !important;
     }
 }
